backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==0.21.1
sqlalchemy==1.4.48
psycopg2-binary==2.9.6
alembic==1.11.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0